I just posted the description for the whole project.
I just posted this description that would give the whole idea about what's the project about. Feel free to have your ideas, but make sure to document every commit you make while you add something to this project.

The objective is to be clear and precise on what you do.